+++ test/unit/fuzz-tools.test.js
@@ -490,6 +490,7 @@ import * as vaizeyV512 from '../../lib/vaizey-v512.js';
 import * as asrsV513 from '../../lib/asrs-v513.js';
 import * as ymrsV514 from '../../lib/ymrs-v514.js';
 import * as simpsonAngusV515 from '../../lib/simpson-angus-v515.js';
+import * as asthmaControlTestV516 from '../../lib/asthma-control-test-v516.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -925,6 +926,7 @@ const MODULES = {
   'asrs-v513.js': asrsV513,
   'ymrs-v514.js': ymrsV514,
   'simpson-angus-v515.js': simpsonAngusV515,
+  'asthma-control-test-v516.js': asthmaControlTestV516,
   'tb-testing.js': tbTesting,
 };